+++ CMakeLists.txt
@@ -37,3 +37,4 @@ include(global_compile_options)
 
 # sub-projects
 add_subdirectory(sdk/core/azure-core)
+add_subdirectory(sdk/platform/http_client/curl) # will work only if BUILD_CURL_TRANSPORT=ON
