+++ sdk/core/azure-core/CMakeLists.txt
@@ -11,6 +11,7 @@ add_library (
   azure-core
   src/http/http
   src/http/request
+  src/http/response
   )
 
 target_include_directories (azure-core PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/inc> $<INSTALL_INTERFACE:include/az_core>)
