+++ sdk/core/azure-core/src/http/response.cpp
@@ -6,12 +6,8 @@
 #include <vector>
 
 #include <http/http.hpp>
-#include <http/response.hpp>
 
 using namespace azure::core::http;
-
-BodyStream& BodyStream::null = *(BodyStream*)NULL;
-BodyBuffer& BodyBuffer::null = *(BodyBuffer*)NULL;
 
 uint16_t const& Response::getStatusCode() { return _statusCode; }
 
@@ -19,6 +15,6 @@ std::string const& Response::getReasonPhrase() { return _reasonPhrase; }
 
 std::map<std::string, std::string> const& Response::getHeaders() { return this->_headers; }
 
-BodyStream& Response::getBodyStream() { return _bodyStream; }
+BodyStream* Response::getBodyStream() { return _bodyStream; }
 
-BodyBuffer& Response::getBodyBuffer() { return _bodyBuffer; }
+BodyBuffer* Response::getBodyBuffer() { return _bodyBuffer; }
