+++ sdk/core/azure-core/test/CMakeLists.txt
@@ -5,15 +5,17 @@ cmake_minimum_required (VERSION 3.12)
 
 include(AddGoogleTest)
 
-project (azure-core-test LANGUAGES CXX)
+set(TARGET_NAME "azure-core-test")
+
+project (${TARGET_NAME} LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 add_executable (
-     azure-core-test
+     ${TARGET_NAME}
      main.cpp 
 )
 
-target_link_libraries(azure-core-test PRIVATE azure-core)
-add_gtest(azure-core-test)
+target_link_libraries(${TARGET_NAME} PRIVATE azure-core)
+add_gtest(${TARGET_NAME})
