+++ sdk/platform/http_client/curl/src/azure_transport_curl.cpp
@@ -1,0 +1,3 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
