+++ sdk/platform/http_client/nohttp/src/azure_transport_nohttp.cpp
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#include <http/http.hpp>
+#include <http/request.hpp>
+#include <http/response.hpp>
+
+using namespace azure::core::http;
+
+// implement send method
+Response Client::send(Request const& request) { throw; }
